requirements.txt
absl-py==0.13.0
asn1crypto==0.24.0
astor==0.8.1
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
cached-property==1.5.2
certifi==2018.1.18
chardet==3.0.4
click==8.0.1
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Flask==2.0.1
gast==0.5.2
google-pasta==0.2.0
grpcio==1.39.0
h5==0.4.1
h5py==2.10.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.6.4
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
JPype1-py3==0.5.5.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
konlpy==0.5.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
netifaces==0.10.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.62.2
tweepy==3.10.0
Twisted==17.9.0
typing-extensions==3.10.0.0
urllib3==1.22
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
zope.interface==4.3.2
